[chore][graph] Split test file
